refactor and  reorder files
